Ajustes de metodos en tipo de relacionado de contribuyente service

+++ contribuyente.microservice/src/main/java/com/nsrtm/contribuyente/microservice/domain/DomicilioRelacionado.java
@@ -47,7 +47,7 @@ public class DomicilioRelacionado extends Auditoria  {
     public Integer viaDistritoId;
 
     @Column(name = "VIA_ID")
-    public Integer viaId;
+    public Long viaId;
 
     @Column(name = "NUMERO_1")
     public Integer numero1;
